+++ nodedb.go
@@ -159,6 +159,20 @@ func (ndb *nodeDB) GetNode(nk []byte) (*Node, error) {
 	if err != nil {
 		return nil, fmt.Errorf("can't get node %v: %v", nk, err)
 	}
+	if buf == nil && !isLegcyNode {
+		// if the node is reformatted by pruning, check against (version, 0)
+		nKey := GetNodeKey(nk)
+		if nKey.nonce == 1 {
+			nodeKey = ndb.nodeKey((&NodeKey{
+				version: nKey.version,
+				nonce:   0,
+			}).GetKey())
+			buf, err = ndb.db.Get(nodeKey)
+			if err != nil {
+				return nil, fmt.Errorf("can't get the reformatted node %v: %v", nk, err)
+			}
+		}
+	}
 	if buf == nil {
 		return nil, fmt.Errorf("Value missing for key %v corresponding to nodeKey %x", nk, nodeKey)
 	}
@@ -190,7 +204,7 @@ func (ndb *nodeDB) GetFastNode(key []byte) (*fastnode.Node, error) {
 	defer ndb.mtx.Unlock()
 
 	if len(key) == 0 {
-		return nil, fmt.Errorf("nodeDB.GetFastNode() requires key, len(key) equals 0")
+		return nil, errors.New("nodeDB.GetFastNode() requires key, len(key) equals 0")
 	}
 
 	if cachedFastNode := ndb.fastNodeCache.Get(key); cachedFastNode != nil {
@@ -331,7 +345,7 @@ func (ndb *nodeDB) shouldForceFastStorageUpgrade() (bool, error) {
 	versions := strings.Split(ndb.storageVersion, fastStorageVersionDelimiter)
 
 	if len(versions) == 2 {
-		latestVersion, err := ndb.getLatestVersion()
+		_, latestVersion, err := ndb.getLatestVersion()
 		if err != nil {
 			// TODO: should be true or false as default? (removed panic here)
 			return false, err
@@ -346,7 +360,7 @@ func (ndb *nodeDB) shouldForceFastStorageUpgrade() (bool, error) {
 // saveFastNodeUnlocked saves a FastNode to disk.
 func (ndb *nodeDB) saveFastNodeUnlocked(node *fastnode.Node, shouldAddToCache bool) error {
 	if node.GetKey() == nil {
-		return fmt.Errorf("cannot have FastNode with a nil value for key")
+		return errors.New("cannot have FastNode with a nil value for key")
 	}
 
 	// Save node bytes to db.
@@ -402,15 +416,54 @@ func (ndb *nodeDB) saveNodeFromPruning(node *Node) error {
 	return ndb.batch.Set(ndb.nodeKey(node.GetKey()), buf.Bytes())
 }
 
+// rootkey cache of two elements, attempting to mimic a direct-mapped cache.
+type rootkeyCache struct {
+	// initial value is set to {-1, -1}, which is an invalid version for a getrootkey call.
+	versions [2]int64
+	rootKeys [2][]byte
+	next     int
+}
+
+func (rkc *rootkeyCache) getRootKey(ndb *nodeDB, version int64) ([]byte, error) {
+	// Check both cache entries
+	for i := 0; i < 2; i++ {
+		if rkc.versions[i] == version {
+			return rkc.rootKeys[i], nil
+		}
+	}
+
+	rootKey, err := ndb.GetRoot(version)
+	if err != nil {
+		return nil, err
+	}
+	rkc.setRootKey(version, rootKey)
+	return rootKey, nil
+}
+
+func (rkc *rootkeyCache) setRootKey(version int64, rootKey []byte) {
+	// Store in next available slot, cycling between 0 and 1
+	rkc.versions[rkc.next] = version
+	rkc.rootKeys[rkc.next] = rootKey
+	rkc.next = (rkc.next + 1) % 2
+}
+
+func newRootkeyCache() *rootkeyCache {
+	return &rootkeyCache{
+		versions: [2]int64{-1, -1},
+		rootKeys: [2][]byte{},
+		next:     0,
+	}
+}
+
 // deleteVersion deletes a tree version from disk.
 // deletes orphans
-func (ndb *nodeDB) deleteVersion(version int64) error {
-	rootKey, err := ndb.GetRoot(version)
+func (ndb *nodeDB) deleteVersion(version int64, cache *rootkeyCache) error {
+	rootKey, err := cache.getRootKey(ndb, version)
 	if err != nil {
 		return err
 	}
 
-	if err := ndb.traverseOrphans(version, version+1, func(orphan *Node) error {
+	if err := ndb.traverseOrphansWithRootkeyCache(cache, version, version+1, func(orphan *Node) error {
 		if orphan.nodeKey.nonce == 0 && !orphan.isLegacy {
 			// if the orphan is a reformatted root, it can be a legacy root
 			// so it should be removed from the pruning process.
@@ -443,7 +496,7 @@ func (ndb *nodeDB) deleteVersion(version int64) error {
 	}
 
 	// check if the version is referred by the next version
-	nextRootKey, err := ndb.GetRoot(version + 1)
+	nextRootKey, err := cache.getRootKey(ndb, version+1)
 	if err != nil {
 		return err
 	}
@@ -525,7 +578,7 @@ func (ndb *nodeDB) deleteLegacyVersions(legacyLatestVersion int64) error {
 
 // DeleteVersionsFrom permanently deletes all tree versions from the given version upwards.
 func (ndb *nodeDB) DeleteVersionsFrom(fromVersion int64) error {
-	latest, err := ndb.getLatestVersion()
+	_, latest, err := ndb.getLatestVersion()
 	if err != nil {
 		return err
 	}
@@ -586,7 +639,7 @@ func (ndb *nodeDB) startPruning() {
 	for {
 		select {
 		case <-ndb.ctx.Done():
-			ndb.done <- struct{}{}
+			close(ndb.done)
 			return
 		default:
 			ndb.mtx.Lock()
@@ -642,7 +695,7 @@ func (ndb *nodeDB) deleteVersionsTo(toVersion int64) error {
 		return err
 	}
 
-	latest, err := ndb.getLatestVersion()
+	_, latest, err := ndb.getLatestVersion()
 	if err != nil {
 		return err
 	}
@@ -670,8 +723,9 @@ func (ndb *nodeDB) deleteVersionsTo(toVersion int64) error {
 		ndb.resetLegacyLatestVersion(-1)
 	}
 
+	rootkeyCache := newRootkeyCache()
 	for version := first; version <= toVersion; version++ {
-		if err := ndb.deleteVersion(version); err != nil {
+		if err := ndb.deleteVersion(version, rootkeyCache); err != nil {
 			return err
 		}
 		ndb.resetFirstVersion(version + 1)
@@ -724,10 +778,11 @@ func (ndb *nodeDB) getFirstVersion() (int64, error) {
 	if itr.Valid() {
 		var version int64
 		legacyRootKeyFormat.Scan(itr.Key(), &version)
+		ndb.resetFirstVersion(version)
 		return version, nil
 	}
 	// Find the first version
-	latestVersion, err := ndb.getLatestVersion()
+	_, latestVersion, err := ndb.getLatestVersion()
 	if err != nil {
 		return 0, err
 	}
@@ -797,13 +852,13 @@ func (ndb *nodeDB) resetLegacyLatestVersion(version int64) {
 	ndb.legacyLatestVersion = version
 }
 
-func (ndb *nodeDB) getLatestVersion() (int64, error) {
+func (ndb *nodeDB) getLatestVersion() (bool, int64, error) {
 	ndb.mtx.Lock()
 	latestVersion := ndb.latestVersion
 	ndb.mtx.Unlock()
 
 	if latestVersion > 0 {
-		return latestVersion, nil
+		return true, latestVersion, nil
 	}
 
 	itr, err := ndb.db.ReverseIterator(
@@ -811,7 +866,7 @@ func (ndb *nodeDB) getLatestVersion() (int64, error) {
 		nodeKeyPrefixFormat.KeyInt64(int64(math.MaxInt64)),
 	)
 	if err != nil {
-		return 0, err
+		return false, 0, err
 	}
 	defer itr.Close()
 
@@ -821,24 +876,25 @@ func (ndb *nodeDB) getLatestVersion() (int64, error) {
 		nodeKeyFormat.Scan(k, &nk)
 		latestVersion = GetNodeKey(nk).version
 		ndb.resetLatestVersion(latestVersion)
-		return latestVersion, nil
+		return true, latestVersion, nil
 	}
 
 	if err := itr.Error(); err != nil {
-		return 0, err
+		return false, 0, err
 	}
 
 	// If there are no versions, try to get the latest version from the legacy format.
 	latestVersion, err = ndb.getLegacyLatestVersion()
 	if err != nil {
-		return 0, err
+		return false, 0, err
 	}
 	if latestVersion > 0 {
 		ndb.resetLatestVersion(latestVersion)
-		return latestVersion, nil
+		return true, latestVersion, nil
 	}
 
-	return 0, nil
+	return false, 0, nil
+	// return -1, nil
 }
 
 func (ndb *nodeDB) resetLatestVersion(version int64) {
@@ -1056,7 +1112,12 @@ func isReferenceRoot(bz []byte) (bool, int) {
 // traverseOrphans traverses orphans which removed by the updates of the curVersion in the prevVersion.
 // NOTE: it is used for both legacy and new nodes.
 func (ndb *nodeDB) traverseOrphans(prevVersion, curVersion int64, fn func(*Node) error) error {
-	curKey, err := ndb.GetRoot(curVersion)
+	cache := newRootkeyCache()
+	return ndb.traverseOrphansWithRootkeyCache(cache, prevVersion, curVersion, fn)
+}
+
+func (ndb *nodeDB) traverseOrphansWithRootkeyCache(cache *rootkeyCache, prevVersion, curVersion int64, fn func(*Node) error) error {
+	curKey, err := cache.getRootKey(ndb, curVersion)
 	if err != nil {
 		return err
 	}
@@ -1066,7 +1127,7 @@ func (ndb *nodeDB) traverseOrphans(prevVersion, curVersion int64, fn func(*Node)
 		return err
 	}
 
-	prevKey, err := ndb.GetRoot(prevVersion)
+	prevKey, err := cache.getRootKey(ndb, prevVersion)
 	if err != nil {
 		return err
 	}
@@ -1105,13 +1166,14 @@ func (ndb *nodeDB) traverseOrphans(prevVersion, curVersion int64, fn func(*Node)
 
 // Close the nodeDB.
 func (ndb *nodeDB) Close() error {
-	ndb.mtx.Lock()
-	defer ndb.mtx.Unlock()
-
 	ndb.cancel()
+
 	if ndb.opts.AsyncPruning {
 		<-ndb.done // wait for the pruning process to finish
 	}
+
+	ndb.mtx.Lock()
+	defer ndb.mtx.Unlock()
 
 	if ndb.batch != nil {
 		if err := ndb.batch.Close(); err != nil {
@@ -1246,7 +1308,7 @@ func (ndb *nodeDB) traverseStateChanges(startVersion, endVersion int64, fn func(
 	if startVersion < firstVersion {
 		startVersion = firstVersion
 	}
-	latestVersion, err := ndb.getLatestVersion()
+	_, latestVersion, err := ndb.getLatestVersion()
 	if err != nil {
 		return err
 	}
@@ -1323,4 +1385,4 @@ func (ndb *nodeDB) String() (string, error) {
 	return "-" + "\n" + buf.String() + "-", nil
 }
 
-var ErrNodeMissingNodeKey = fmt.Errorf("node does not have a nodeKey")
+var ErrNodeMissingNodeKey = errors.New("node does not have a nodeKey")